+++ tests/typed_data_obj.rs
@@ -41,9 +41,9 @@ impl Line {
 }
 
 impl DataTypeFunctions for Line {
-    fn mark(&self) {
-        gc::mark(self.start());
-        gc::mark(self.end());
+    fn mark(&self, marker: &gc::Marker) {
+        marker.mark(self.start);
+        marker.mark(self.end);
     }
 }